refactor: change folder of default data
