+++ src/pageComponents/HomePage/defaultData/footerSocials.ts
